chore(stories): add actions to storybook
