--- storybook-old/app/builtin/init-stories.js
@@ -1,55 +1,0 @@
-import initAnalyticsStories from './stories/analytics.stories';
-import initBreadcrumbStories from './stories/breadcrumb.stories.js';
-import initClearRefinementsStories from './stories/clear-refinements.stories';
-import initCurrentRefinementsStories from './stories/current-refinements.stories';
-import initGeoSearch from './stories/geo-search.stories';
-import initHierarchicalMenu from './stories/hierarchical-menu.stories';
-import initHitsStories from './stories/hits.stories';
-import initHitsPerPageStories from './stories/hits-per-page.stories';
-import initInfiniteHitsStories from './stories/infinite-hits.stories';
-import initInstantSearchStories from './stories/instantsearch.stories';
-import initMenuStories from './stories/menu.stories';
-import initMenuSelectStories from './stories/menu-select.stories';
-import initNumericMenuStories from './stories/numeric-menu.stories';
-import initPaginationStories from './stories/pagination.stories';
-import initRangeInputStories from './stories/range-input.stories.js';
-import initRangeSliderStories from './stories/range-slider.stories';
-import initRefinementListStories from './stories/refinement-list.stories';
-import initReloadStories from './stories/reload.stories';
-import initSearchBoxStories from './stories/search-box.stories';
-import initSortByStories from './stories/sort-by.stories';
-import initRatingMenuStories from './stories/rating-menu.stories';
-import initStatsStories from './stories/stats.stories';
-import initToggleStories from './stories/toggleRefinement.stories';
-import initConfigureStories from './stories/configure.stories';
-import initPoweredByStories from './stories/powered-by.stories';
-import initPanelStories from './stories/panel.stories';
-
-export default () => {
-  initAnalyticsStories();
-  initBreadcrumbStories();
-  initClearRefinementsStories();
-  initCurrentRefinementsStories();
-  initGeoSearch();
-  initHierarchicalMenu();
-  initHitsStories();
-  initHitsPerPageStories();
-  initInfiniteHitsStories();
-  initInstantSearchStories();
-  initMenuStories();
-  initMenuSelectStories();
-  initNumericMenuStories();
-  initPaginationStories();
-  initRangeInputStories();
-  initRangeSliderStories();
-  initRefinementListStories();
-  initReloadStories();
-  initSearchBoxStories();
-  initSortByStories();
-  initStatsStories();
-  initRatingMenuStories();
-  initToggleStories();
-  initConfigureStories();
-  initPoweredByStories();
-  initPanelStories();
-};
