--- storybook-old/app/instantsearch.js
@@ -1,3 +1,0 @@
-import instantsearch from '../../src/index';
-
-export default instantsearch;
