--- storybook-old/app/utils/all-items.html
@@ -1,10 +1,0 @@
-<table class="table table-striped table-condensed">
-  <thead>
-    <tr><th>name</th><th>brand</th><th>price</th></tr>
-  </thead>
-  <tbody>
-  {{#hits}}
-    <tr><td><a href="#hit-{{objectID}}">{{name}}</a></td><td>{{brand}}</td><td>{{price}}</td></tr>
-  {{/hits}}
-  </tbody>
-</table>
