--- storybook-old/app/utils/item.html
@@ -1,18 +1,0 @@
-<div class="hit" id="hit-{{objectID}}">
-  <div class="hit-picture">
-    <img src="{{image}}">
-  </div>
-  <div class="hit-content">
-    <div>
-      <span>{{{_highlightResult.name.value}}}</span>
-      <span>{{price_formatted}}</span>
-      <span>{{rating}} stars</span>
-    </div>
-    <div class="hit-type">
-      {{{_highlightResult.type.value}}}
-    </div>
-    <div class="hit-description">
-      {{{_highlightResult.description.value}}}
-    </div>
-  </div>
-</div>
