--- storybook-old/app/utils/no-results.html
@@ -1,1 +1,0 @@
-No matching products, try another search.
