--- storybook-old/template.html
@@ -1,17 +1,0 @@
-<!doctype html>
-<html lang="en">
-
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=Edge">
-  <title>Instant search demo built with instantsearch.js</title>
-
-  <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/instantsearch.css@7.1.0/themes/algolia.min.css">
-
-  <script src="https://cdn.polyfill.io/v2/polyfill.min.js?features=default,Array.from"></script>
-</head>
-
-<body>
-</body>
-
-</html>
